user predicate/requirements.txt
# requirements.txt
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
xgboost==1.7.5
matplotlib==3.7.1
seaborn==0.12.2
PyYAML==6.0
